- refactor all iterators and enumerables into separate files
